Add methods to validate and parse task strings in Task class
The Task class now includes two new methods: isValidTaskString and
parseTaskString. The isValidTaskString method checks if a given string
is a valid representation of a task, while the parseTaskString method
converts a valid task string into a Task object.

These methods are necessary for the storage of tasks. When tasks are
stored as strings in a JSON file, these methods will be used to validate
the task strings and convert them back into Task objects when the data
is loaded. This ensures that the tasks are stored and loaded correctly,
and any invalid task strings are detected and handled appropriately.